Fixed signage bug when computing time comparison
The value returned by GetZoneTicks is unsigned, and so sometimes wraps around to its upper bound, causing diff to be a very large number. Static casted these to ints to prevent that.

+++ mp/src/game/client/momentum/ui/HUD/hud_mapfinished.cpp
@@ -514,9 +514,9 @@ void CHudMapFinishedDialog::SetComparisonForPage(int iPage)
         return;
     }
 
-    int runTicks = iPage == 0 ? m_pRunData->m_iRunTime : m_pRunStats->GetZoneTicks(iPage);
+    int runTicks = iPage == 0 ? m_pRunData->m_iRunTime : static_cast<int>(m_pRunStats->GetZoneTicks(iPage));
 
-    float diff = float(runTicks - m_pPbRunStats.GetZoneTicks(iPage)) * m_pRunData->m_flTickRate;
+    float diff = static_cast<float>(runTicks - static_cast<int>(m_pPbRunStats.GetZoneTicks(iPage))) * m_pRunData->m_flTickRate;
 
     char diffString[BUFSIZETIME];
     Color diffColor;
